Invalid source code layout for `eth_account`. Handling logic moved to separate method
